+++ python/evebus/rpc/__init__.py
@@ -1,0 +1,8 @@
+"""
+pyevebus RPC 客户端 SDK
+
+    from evebus.rpc import RPCClient
+"""
+from .client import RPCClient, RPCError, DEFAULT_URL
+
+__all__ = ["RPCClient", "RPCError", "DEFAULT_URL"]